==> markdown/4.analysis-den.Rmd <==
---
title: "4.analysis-den"
output: html_document
---

# Setup  
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

# Libraries
library(tidyverse)
library(janitor)
library(ggthemes)
library(readxl)
library(sf)
library(USAboundaries)
library(partykit)
library(car)
library(emmeans)
library(multcomp)

source("../markdown/00_functions.R")

```

```{r theme}
theme_set(
  theme_excel()+
    theme(panel.grid = element_line(size=.1))
)
```

```{r den_w}
den_w <- read_csv("../data/den_w.csv") %>%
  mutate_if(is.character, as.factor) %>%
  dplyr::select(-wgt) %>%
  rename(LRR=LRRden, wgt=weight)

```

```{r den NA?}
den_w %>%
  dplyr::select(LRR) %>%
  summary

# no NAs
```

```{r sd per study?}
den_w %>%
  group_by(study) %>%
  tally
```

# Analysis  
```{r full model}
dres <- RunModelNoModsFun(mydata = den_w)

dres
```

```{r individual studies}
den_studies <- den_w %>%
  group_by(study, pub_reference) %>%
  nest() %>%
  mutate(dres_study=map(data,
                        ~RunModelOneStudyFun(
                          mydata = .x
                        )))


den_studies %>%
  unnest(dres_study) %>%
    dplyr::select(pub_reference, 
           estimate, std.error, p.value, 
           cilo, cihi) %>% 
  write_csv("../output/den-ss_effect-size-by-study.csv")

```

```{r loo by study}
dreslo <- dres %>% 
  mutate(studylo = NA)

den_stud <- unique(den_w$study)

for (i in 1:length(den_stud)) {
  
  # trouble
  #i <- 6
  lo <- den_stud[i]
  d.tmp <- den_w %>% 
    filter(study != lo)
  
  dres.tmp <- RunModelNoModsFun(mydata = d.tmp) %>% 
    mutate(studylo = lo)

  dreslo <- bind_rows(dreslo, dres.tmp)

}

dreslo  

# Exporting
dreslo %>% 
  mutate_if(is.numeric, round, 2) %>% 
  dplyr::select(-pavg, -plo, -phi) %>% 
  write_csv("../output/den-ss_loo-overall.csv")

# Plotting
dreslo

```
the first row on `dreslo` is the full model.



```{r covariates with cc biomass}
options(contrasts = c("contr.sum", "contr.poly"))

cc_cov <- den_w %>%
  dplyr::select(study, cc_bio_Mgha,
                lat,
                cc_type,
                cc_term.METH2,
                tillage_sys,
                #OM_pct too many NAs
                ) %>%
  gather(var, val, -study, -cc_bio_Mgha) %>%
  group_by(var) %>%
  nest() %>%
  mutate(data_w=map2(data, var,
                    ~.x %>%
                      mutate(var=.y,
                        val=ifelse(var == "lat", 
                                   as.numeric(val),
                        
                                  as.character(as.factor(val))
                        ))
                    )) %>%
  mutate(mod=map(data_w,
                 ~lmer(cc_bio_Mgha ~ val + (1|study),
                       data = .x)) 
                 ) %>%
  mutate(anova=map(mod,
                     ~Anova(.x, type=3) %>%
                       as.data.frame() %>%
                     rownames_to_column(var = "source")
                     ))


# Which of the covariates above were significant?
cc_cov %>%
  unnest(anova) %>%
  filter(source == "val") %>%
  filter(`Pr(>Chisq)`<0.1)
```

```{r cctype vs cc_biomass}
letter <- cc_cov %>%
  filter(var == "cc_type") %>%
  mutate(means=map2(mod,var,
                   ~emmeans(.x, ~val
                            )
                   )) %>%
  mutate(cld=map(means,
                 ~cld(.x,
                      reversed=T,
                      adjust="none",
                      Letters=letters) %>%
                   as.data.frame() %>%
                   mutate(letter=trimws(.group))
                 )) %>%
  unnest(cld) %>%
  rename(cc_type=val)

letter
  
den_w %>%
  ggplot(aes(y=cc_bio_Mgha,
             x=cc_type))+
  geom_boxplot()+
  geom_point(data = letter,
             aes(y=emmean),
             color="purple",
             shape=4,
             size=2)+
  geom_label(data = letter, 
             label.padding = unit(.1, "lines"),
             aes(y=emmean+1,
                 label=letter),
             color="blue")

```

```{r LRR ~ cc_type*cc_biomass}

den_w %>%
  ggplot(aes(x=cc_bio_Mgha,
             y=LRR,
              color=cc_type))+
  geom_point()+
  geom_smooth(method="lm")

```

```{r contrasts}
themods <- c("cc_type", 
             "cc_term.METH2", 
             "tillage_sys")

denmc <- RunModelModsContrastFun(mydata = den_w, 
                                 mymod = themods[1])

for (i in 2:length(themods)){
  tmp <- RunModelModsContrastFun(mydata = den_w, 
                                 mymod = themods[i])
  suppressWarnings(
  denmc <- bind_rows(denmc, tmp)
  )
}

```

```{r continuous modifiers}
# jsut for reference
thecontmods <- c("pub_year", 
                 "lat", 
                 "long",
                 "cc_bio_Mgha")


#--density
denmcont <- RunModelContModsFun(mydata = den_w, 
                                mymod = thecontmods[1])

for (i in 2:length(thecontmods)){
  tmp <- RunModelContModsFun(mydata = den_w, 
                             mymod = thecontmods[i])
  denmcont <- bind_rows(denmcont, tmp)
}

```


```{r}
contmodres <- denmcont

contmodres %>% 
  write_csv("../output/den-ss_contmods.csv")
```

# LRR vs cc_bio  
```{r}
den_w %>%
  drop_na(LRR, cc_bio_Mgha) %>%
  ggplot(aes(y=LRR, x=cc_bio_Mgha))+
  geom_point()+
  geom_smooth(method="lm")
```
## Linear
```{r den_lm}
den_lm <- lm(LRR ~ cc_bio_Mgha, 
     data = den_w %>%
       drop_na(LRR, cc_bio_Mgha)
     )

den_lm 

AICcmodavg::AICc(den_lm)
BIC(den_lm)
```


## LP
```{r f.lrp}
f.lrp <- function(x, a, b, tx) {
  ifelse(x > tx, 
         a + b * tx, 
         a + b * x)}
```

```{r linp_mod}
linp_mod <- nls(LRR ~ f.lrp(cc_bio_Mgha, a, b, tx),
                data = den_w,
                start = list(a = -0.1,
                              b = -.1,
                              tx = 3)
                )

linp_mod

summary(linp_mod)

library(Metrics)

BIC(linp_mod)
```

```{r lp res}
plot(linp_mod)
```

```{r den_w_aug}
den_w_aug <- augment(linp_mod, data=den_w) %>%
  mutate(.stdresid=resid(linp_mod, type="pearson", scaled=T))
  
  
den_w_aug
```

```{r lp final plot}
ggplot(den_w_aug, aes(x=cc_bio_Mgha, y=LRR))+
  geom_point()+
  geom_line(aes(y=.fitted), color="purple")
```

## QP  
```{r f.qp}
# a = intercept
# b = slope
# c = quadratic term (curvy bit)
# tx = join point = break point = critical concentration

f.qp <- function(x, a, b, tx) {
    c <- -0.5 * b / tx
    if_else(condition = x < tx,
            true  = a + (b * x) + (c * x * x),
            false = a + (b * tx) + (c * tx * tx))
}
```

```{r qp_mod}
qp_mod <- nls(LRR ~ f.qp(cc_bio_Mgha, a, b, tx),
                data = den_w,
                start = list(a = -0.1,
                              b = -.1,
                              tx = 3)
                )

qp_mod

summary(qp_mod)

BIC(qp_mod)
```

```{r qp res}
plot(qp_mod)
```

```{r den_w_aug}
den_w_qp_aug <- augment(qp_mod, data=den_w) %>%
  mutate(.stdresid=resid(linp_mod, type="pearson", scaled=T))
  
  
den_w_qp_aug

```

```{r qp final plot}
ggplot(den_w_qp_aug, aes(x=cc_bio_Mgha, y=LRR))+
  geom_point()+
  #geom_point(aes(y=.fixed), color="purple")+
  geom_line(aes(y=.fitted), color="purple")
```

## Final - lm
linear was lowest BIC, so decided to keep it.
Finding 75% reduction:

```{r}
1-.25
log(.5)

exp(-0.6931472)
```
For 50% reduction in weed density, need to predict x value where LRR=-0.6931472

```{r cc_bio_50}
cc_bio_50 <- (-0.6931472 - coef(den_lm)[[1]])/coef(den_lm)[[2]]

cc_bio_50

```



# Meeting with David on June 3th

Defense 22-25 August
Moving end of July
Submit meta-analysis to Nick by July 15th (4-th to 15th working hard)  


- Discussion about weight. His weight is correct, used sample_size instead of reps.

- Do CC suppress weed biomass or density?  (histograms, done)
- How much? 
- Why?  
- What about yield effect?  

- Histograms
- Maps
- Tree for density
